apply useMemo to throttledSetMarkers ...
AS-IS:
throttledSetMarkers becomes new function at every rendering
-> trigger Mainviewer's useEffect which depends on throttledSetMarkers
-> the useEffect calls throttledSetMarkers
-> markersRef.current becomes new object
-> Because markersRef.current becomes new object at every rendering,
AxisCanvas.draw should compare markers' items with the previous one
to determine drawing is needed.

TO-BE:
throttledSetMarkers doesn't change at all.
throttledSetMarkers is called only when it is needed
AxisCanvas.draw will draw axis only if markersRef.current is a new obj.