setLensParams for fit canvas is run by useLayoutEffect, not useEffect ...
Because setStates in useEffect fire re-rendering, pxPerSec is updated in the *next* re-rendering after width changes.
If canvasIsFit, ImgCanvas and Overview are re-rendered twice when width changes (1. re-rendering with changed width, 2. re-rendering with changed pxPerSec). It cause jittering of overview lens, and amplifying jittering of ImgCanvas.
useLayoutEffect can solve this problem.