refactor backend to prevent occupying rayon threads to wait for write locks ...
AS-IS: When users change sth so quickly, it eventually blocks all rayon threads -> data parallel calc cannot run

To resolve that, a system thread is spawned and handles TRACK_LIST, TM write locks exclusively.
By this, requests that uses write locks run serially, but it's not a big problem because write locks can exist one at a time anyway.
If we don't spawn a thread and handle write locks in the main thread, it blocks UI. So we need at least one thread.
tokio blocking threads are slow because they're created on-demand way (roughly speaking).
So I decided to spawn a normal system thread.